getTypeIdの返り値を修正

+++ src/TAGString.java
@@ -42,7 +42,7 @@ public class TAGString extends TAGComponent{
 	}
 
 	@Override
-	public int getTypeId() {
+	public byte getTypeId() {
 		return 8;
 	}
